+++ core/standalone-decorator/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.36.3](https://github.com/eggjs/tegg/compare/v3.36.2...v3.36.3) (2024-04-10)
+
+**Note:** Version bump only for package @eggjs/standalone-decorator
+
+
+
+
+
+## [3.36.2](https://github.com/eggjs/tegg/compare/v3.36.1...v3.36.2) (2024-04-08)
+
+**Note:** Version bump only for package @eggjs/standalone-decorator
+
+
+
+
+
+## [3.36.1](https://github.com/eggjs/tegg/compare/v3.36.0...v3.36.1) (2024-04-07)
+
+**Note:** Version bump only for package @eggjs/standalone-decorator
+
+
+
+
+
 # [3.36.0](https://github.com/eggjs/tegg/compare/v3.35.1...v3.36.0) (2024-04-02)
 
 **Note:** Version bump only for package @eggjs/standalone-decorator